requirements.txt
Click==7.0
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
progress-bar==8
python-dateutil==2.8.1
pytz==2019.3
scenedetect==0.5.1.1
six==1.13.0
srt==3.0.0
timecode==1.2.0
